move pages tests
